requirements.txt
python-telegram-bot==20.7
httpx>=0.25.0
python-dotenv==1.0.0
cryptography==41.0.7
pandas==2.1.3
python-dateutil==2.8.2
aiohttp==3.9.1
APScheduler==3.10.4